ceph: improve gate builds by reducing warn states
Since rgw comes up and there is a consul test to create swift buckets
and delete, buckets index pool gets skewed a bit and ceph health goes
into a warn state, which increases the amount of puppet runs to bring
the state to heatlhy. Overriding this ratio for gate builds.

Reported-by: Anshu Prateek <anshu.prateek@ril.com>
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@gmail.com>