cinder: handle concurrent db connections better
Increase the value of sqlalchemy's pool size, which allows more requests
to be taken, avoids cinder errroring out with a 500. This has been seen
to affect cases whenever an action such as creation/deletion of say 100
volumes or 100 backups etc. For some more info refer to an upstream
bug: https://bugs.launchpad.net/cinder/+bug/1224608

This should fix a couple of current bugs against block storage:
Fixes: https://bugs.launchpad.net/jio/+bug/1470472 &
https://bugs.launchpad.net/jio/+bug/1467915

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@ril.com>